+++ src/handlers/focus-handler.js
@@ -27,3 +27,20 @@ export const addBlurHandler = (element) => {
     });
   };
 };
+
+export const saveFocusedElement = (name) => {
+  sessionStorage.setItem("lastFocused", name);
+};
+
+export const restoreFocus = () => {
+  const lastFocusedName = sessionStorage.getItem("lastFocused");
+  if (lastFocusedName) {
+    const element = document.querySelector(
+      `[lastFocused="${lastFocusedName}"]`
+    );
+    if (element) {
+      addFocus(element);
+      sessionStorage.removeItem("lastFocused");
+    }
+  }
+};
